+++ frontend/src/hooks/useTransactionSummary.js
@@ -1,0 +1,26 @@
+import { useRecoilValue } from 'recoil';
+
+import { getTransactionsInDateState } from '../recoil/transaction/selector';
+
+import { calculateIncome, calculateExpenditure } from '../utils/common/calculate-cost';
+
+export const useTransactionSummary = () => {
+    const rawTransactions = useRecoilValue(getTransactionsInDateState);
+
+    const transactionCount = Object.values(rawTransactions).reduce(
+        (acc, transaction) => acc + transaction.length,
+        0,
+    );
+
+    let monthIncome = 0;
+    Object.values(rawTransactions).forEach((rawTransaction) => {
+        monthIncome += calculateIncome(rawTransaction);
+    });
+
+    let monthExpenditure = 0;
+    Object.values(rawTransactions).forEach((rawTransaction) => {
+        monthExpenditure += calculateExpenditure(rawTransaction);
+    });
+
+    return [transactionCount, monthIncome, monthExpenditure];
+};
